pyproject.toml
[project]
name = "ai-dj"
version = "0.0.1"
description = "AI DJ — MERT-based music embedding and path-planning player"
requires-python = ">=3.10,<3.12"
dependencies = [
    "torch==2.9.1+rocm6.4",
    "torchaudio==2.9.1+rocm6.4",
    "pytorch-triton-rocm",
    "numpy",
    "transformers>=4.40",
    "librosa>=0.10",
    "soundfile",
    "qdrant-client>=1.17,<2",
    "mutagen>=1.47",
    "tqdm>=4.66",
    "umap-learn>=0.5",
    "PySide6>=6.7",
    "pyqtgraph>=0.13",
    "python-mpv>=1.0",
]

[project.scripts]
ai-dj = "ai_dj.gui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_dj"]

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-rocm" }
torchaudio = { index = "pytorch-rocm" }
pytorch-triton-rocm = { index = "pytorch-rocm" }
